Move app.py into sc_flask module.
This touches quite a few files, but is a fairly simple change. Again we're moving a script into the sc_flask module. I've had to move the wsgi.py file as several commands won't work unless uwsgi can find either app.py or wsgi.py

Issue #3564
