+++ server/server/api/tests/test_endpoints.py
@@ -1,7 +1,7 @@
 import pytest
 import json
 
-from app import api, app
+from sc_flask.app import app
 
 
 class TestHomePageData:
