+++ server/server/wsgi.py
@@ -4,7 +4,7 @@ import sys
 dir_path = os.path.dirname(os.path.realpath(__file__))
 sys.path.append(dir_path)
 
-from app import app  # isort:skip
+from sc_flask.app import app  # isort:skip
 
 
 if __name__ == '__main__':
