+++ server/uwsgi.ini
@@ -1,6 +1,6 @@
 [uwsgi]
 socket = /opt/sc/sockets/uwsgi.sock
-module =  sc_flask.wsgi:app
+module =  wsgi:app
 
 master = true
 processes = 4
